requirements.txt
alembic==1.5.8
asn1crypto==0.24.0
Beaker==1.10.0
cached-property==1.5.2
certifi==2021.5.30
cftime==1.3.1
chardet==3.0.4
click==7.1.2
coards==1.0.5
cryptography==3.3.2
enum34==1.1.6
Flask==2.0.3
funcsigs==1.0.2
# GDAL==3.0.4
Genshi==0.7.5
GeoAlchemy==0.7.3
GeoAlchemy2==0.13.3
gitdb2==2.0.6
GitPython==2.1.15
# h5py==2.7.1 - pre-installed in docker
idna==2.6
ipaddress==1.0.17
itsdangerous==2.1.2
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
Mako==1.1.4
MarkupSafe==2.0
modelmeta==1.0.1
nchelpers==5.5.7
netCDF4==1.3.0
numpy==1.21
olefile==0.46
openid2rp==1.12
Paste==3.7.1
pdp_util==3.0.0
Pillow==5.4.1
ply==3.10
psycopg2==2.9.6
pupynere-pdp==2.0.1
PyCDS==4.5.0
PyCRS==1.0.2
pycrypto==2.6.1
Pydap==3.2.2
pydap_extras
# pygobject==3.26.1 - not available, lack doesn't seem to matter
pystache==0.5.4
python-dateutil==2.6.1
python-editor==1.0.4
pytz==2018.9
PyYAML==5.4.1
requests==2.31.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.11.0
slimit==0.8.1
smmap==3.0.5
smmap2==3.0.1
SQLAlchemy==1.4.15 #added for pydap-extras
sqlparse==0.3.1
static==1.1.1
urllib3==1.26.5
WebOb==1.8.7
Werkzeug==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
